reorganize the code into individual files
